+++ contracts/Omni-Chain_DApps/OminiToken/interface/IOmniToken.sol
@@ -5,13 +5,13 @@ interface IOmniToken {
     function mint(address toAddress, uint256 amount) external;
 
     function bridgeTransfer(
-        uint64 destChainId,
+        uint16 destChainId,
         address receiver,
         uint256 amount
     ) external payable;
 
     function fetchOmniTokenTransferFee(
-        uint64[] calldata destChainId,
+        uint16[] calldata destChainId,
         address[] calldata receiver,
         uint256[] calldata amount
     ) external view returns (uint256);
